+++ lib/gds_api/external_link_tracker.rb
@@ -1,7 +1,0 @@
-require_relative 'base'
-
-class GdsApi::ExternalLinkTracker < GdsApi::Base
-  def add_external_link(url)
-    put_json!("#{endpoint}/url?url=#{CGI.escape(url)}", {})
-  end
-end